More instances of C++ type-anal-ness. Also changing ri->ftol to Q_ftol

+++ codemp/rd-rend2/tr_flares.cpp
@@ -346,7 +346,7 @@ void RB_RenderFlare( flare_t *f ) {
  * The coefficient flareCoeff will determine the falloff speed with increasing distance.
  */
 
-	factor = distance + size * sqrt(flareCoeff);
+	factor = distance + size * sqrt((double)flareCoeff);
 	
 	intensity = flareCoeff * size * size / (factor * factor);
